src/main/webapp/app/entities/niveau/update/niveau-update.component.spec.ts
jest.mock('@angular/router');

import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { FormBuilder } from '@angular/forms';
import { ActivatedRoute } from '@angular/router';
import { of, Subject } from 'rxjs';

import { NiveauService } from '../service/niveau.service';
import { INiveau, Niveau } from '../niveau.model';

import { NiveauUpdateComponent } from './niveau-update.component';

describe('Component Tests', () => {
  describe('Niveau Management Update Component', () => {
    let comp: NiveauUpdateComponent;
    let fixture: ComponentFixture<NiveauUpdateComponent>;
    let activatedRoute: ActivatedRoute;
    let niveauService: NiveauService;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [NiveauUpdateComponent],
        providers: [FormBuilder, ActivatedRoute],
      })
        .overrideTemplate(NiveauUpdateComponent, '')
        .compileComponents();

      fixture = TestBed.createComponent(NiveauUpdateComponent);
      activatedRoute = TestBed.inject(ActivatedRoute);
      niveauService = TestBed.inject(NiveauService);

      comp = fixture.componentInstance;
    });

    describe('ngOnInit', () => {
      it('Should update editForm', () => {
        const niveau: INiveau = { id: 456 };

        activatedRoute.data = of({ niveau });
        comp.ngOnInit();

        expect(comp.editForm.value).toEqual(expect.objectContaining(niveau));
      });
    });

    describe('save', () => {
      it('Should call update service on save for existing entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Niveau>>();
        const niveau = { id: 123 };
        jest.spyOn(niveauService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ niveau });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: niveau }));
        saveSubject.complete();

        // THEN
        expect(comp.previousState).toHaveBeenCalled();
        expect(niveauService.update).toHaveBeenCalledWith(niveau);
        expect(comp.isSaving).toEqual(false);
      });

      it('Should call create service on save for new entity', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Niveau>>();
        const niveau = new Niveau();
        jest.spyOn(niveauService, 'create').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ niveau });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.next(new HttpResponse({ body: niveau }));
        saveSubject.complete();

        // THEN
        expect(niveauService.create).toHaveBeenCalledWith(niveau);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).toHaveBeenCalled();
      });

      it('Should set isSaving to false on error', () => {
        // GIVEN
        const saveSubject = new Subject<HttpResponse<Niveau>>();
        const niveau = { id: 123 };
        jest.spyOn(niveauService, 'update').mockReturnValue(saveSubject);
        jest.spyOn(comp, 'previousState');
        activatedRoute.data = of({ niveau });
        comp.ngOnInit();

        // WHEN
        comp.save();
        expect(comp.isSaving).toEqual(true);
        saveSubject.error('This is an error!');

        // THEN
        expect(niveauService.update).toHaveBeenCalledWith(niveau);
        expect(comp.isSaving).toEqual(false);
        expect(comp.previousState).not.toHaveBeenCalled();
      });
    });
  });
});
